fix labels loading time
