+++ package.json
@@ -35,7 +35,6 @@
     "react-redux": "^7.2.6",
     "react-responsive": "^9.0.0-beta.6",
     "react-router-dom": "^6.2.1",
-    "react-window": "^1.8.7",
     "redux-thunk": "^2.4.1",
     "slate": "^0.73.1",
     "slate-history": "^0.66.0",
