+++ src/components/DataView/ValueTags.js
@@ -12,6 +12,8 @@ import styled from "@emotion/styled";
 const ValueTagsStyled = styled.div`
     color: ${(props) => props.theme.textPrimary};
     flex: 1;
+    z-index: 10000;
+    display: flex;
     &:hover {
         background: ${(props) => props.theme.widgetContainer};
     }
